Removed rotation about x-axis.
This rotation was introduced early on the development process and has
only made development painful. The only benefit it had was that it put
the maps for the 2D systems on the ground plane but the math is easier
without the rotation. I will likely reintroduce it as a real rotation
for debugging purposes.